addMonths{Checked}: Ensure type safety and deduplicate
Until now we've gotten away with not properly checking which actual
month is new_month and which actual day of month is the new day of
month. Now that DayOfMonth does proper validity checking, we need to fix
this. Semantically nothing changed, this wasn't a bug per se, but I like
my type safety and that is now satisfied.

Additionally addMonth now simply calls addMonthChecked. This
implementation is too complicated to have twice in the codebase, even if
it's next to eachother.