TZString: deal with out of month occurances
Accodring to spec, if occurance 5 is outside of the month, clamp to the
last occrance in month.

We extend this principle to clamp any out of bounds value to the last
occurance in month. This deals with the possible case where occurance 5
is accessed in february, which can be up to 2 weeks after the month's
end.